# use bison cpp API to support non-trivial types