… symengine/parser/parser.tab.cc
… symengine/parser/parser.tab.hh
… symengine/parser/parser_stype.h
… symengine/parser/tokenizer.cpp