- remove no longer needed `YYSTYPE` struct
- resolves #1804